Fix unit test compile errors
- HdrSupportTest: use Display.HDR_TYPE_* literal values (mockable android.jar lacks constants)
- InverseTonemapTest: kotlin.math.pow is an extension, use 2.0.pow(1.1)
- RenderCoreTest: JUnit has no assertEquals(int,int,int) delta overload

+++ app/src/test/kotlin/io/github/gordonlu/lumadepth/RenderCoreTest.kt
@@ -339,9 +339,9 @@ class RenderCoreTest {
             val r = (v shr 16) and 0xFF
             val g = (v shr 8) and 0xFF
             val b = v and 0xFF
-            assertEquals(128, r, 1)
-            assertEquals(128, g, 1)
-            assertEquals(128, b, 1)
+            assertEquals(128f, r.toFloat(), 1f)
+            assertEquals(128f, g.toFloat(), 1f)
+            assertEquals(128f, b.toFloat(), 1f)
         }
     }
